complete: chapter 14, implemented debug print of bytecode

+++ include/debug.hpp
@@ -6,4 +6,4 @@
 
 auto disassemble_chunk(const Chunk &chunk, const std::string &name) -> void;
 
-auto disassemble_instruction(const Chunk &chunk, size_t offset) -> size_t;
+auto disassemble_instruction(const Chunk &chunk, int offset) -> int;
